[refactor] Redis 정리
