+++ src/main/java/com/project/dogwalker/common/config/RedisConfig.java
@@ -2,7 +2,7 @@ package com.project.dogwalker.common.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.project.dogwalker.common.service.CoordinateDeserializer;
+import com.project.dogwalker.common.service.redis.CoordinateDeserializer;
 import lombok.RequiredArgsConstructor;
 import org.locationtech.jts.geom.Coordinate;
 import org.redisson.Redisson;
@@ -14,7 +14,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.RedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
@@ -58,14 +57,5 @@ public class RedisConfig {
     redisTemplate.setValueSerializer(new StringRedisSerializer());
     return redisTemplate;
   }
-  @Bean
-  public StringRedisTemplate stringRedisTemplate(){
-    StringRedisTemplate stringRedisTemplate=new StringRedisTemplate();
-    stringRedisTemplate.setKeySerializer(new StringRedisSerializer());
-    stringRedisTemplate.setValueSerializer(new StringRedisSerializer());
-    stringRedisTemplate.setConnectionFactory(redisConnectionFactory());
-    return stringRedisTemplate;
-  }
-
 
 }
